refactor(memory-bridge): shrink to <1000 lines via facade + stub parity
Split memory-bridge.ts (1454 -> 650 lines) into three files:
- bridge-core.ts (new, 278 lines): registry singleton, withDb, cosineSim,
  generateId, vector-stats cache, MEMORY_ENTRIES_DDL.
- bridge-entries.ts (new, 530 lines): sql.js memory_entries CRUD + BM25 +
  cache/guard/attestation helpers. makeEntryCacheKey deduped across 3 sites.
- memory-bridge.ts (now 650 lines): thin facade; re-exports + controller-op
  wrappers for hierarchical/consolidate/batch/context/session/HNSW/etc.

Stub parity: createTieredMemoryStub() replaced by HierarchicalMemoryStub
class implementing the full async store/recall/promote/forget/getStats/count
surface. Eliminates 4 `typeof hm.promote === 'function'` duck-type sites in
the bridge (were at memory-bridge.ts:1238,1257,1280,1354).

Also fixes a subpath-traversal bug in getDbPath: resolved.startsWith(cwd)
collided on `moflo-backup` vs `moflo`; now uses path.relative.

Try-block count in memory-bridge.ts: 35 -> 14 (stretch goal <30).

Closes #492
